requirements.txt
langchain
openai
supabase
google-auth-oauthlib
python-dotenv
